packages/macos/codegrey-agent-macos/requirements.txt
# CodeGrey AI SOC Platform - macOS Client Agent Dependencies

# Core dependencies
requests>=2.31.0
psutil>=5.9.0
pyyaml>=6.0
aiohttp>=3.9.0

# Container management
docker>=7.0.0

# Logging and utilities
loguru>=0.7.0
watchdog>=3.0.0

# macOS-specific
pyobjc-core>=9.0; sys_platform == "darwin"
pyobjc-framework-Cocoa>=9.0; sys_platform == "darwin"
